Add files via upload
This was a simple "mini-project" undertaken in 2020. The aim was to use a publicly available dataset, structure some question(s) and convey conclusions via a PowerPoint presentation. Analysis was done using python, NumPy, Pandas, and Matplotlib.

I've uploaded it in PDF format because the powerpoint presentation is attached at the end of the jupyter notebook